:ok_hand: Refs. #614 -- PR feedback
* Moved exceptions from service to exceptions module
* Updated docstring of submission complete action to describe session
  handling
* Added nuance to logger statements
