+++ src/openforms/appointments/service.py
@@ -7,25 +7,13 @@ outside of this app.
 """
 from openforms.submissions.models import Submission
 
+from .exceptions import AppointmentRegistrationFailed, AppointmentUpdateFailed
+
 __all__ = [
     "AppointmentRegistrationFailed",
     "AppointmentUpdateFailed",
     "register_appointment",
 ]
-
-
-class AppointmentInteractionFailed(Exception):
-    def __init__(self, *args, **kwargs):
-        self.should_retry = kwargs.pop("should_retry", False)
-        super().__init__(*args, **kwargs)
-
-
-class AppointmentRegistrationFailed(AppointmentInteractionFailed):
-    pass
-
-
-class AppointmentUpdateFailed(AppointmentInteractionFailed):
-    pass
 
 
 def register_appointment(submission: Submission) -> None:
